+++ extensions/homeassistant/src/components/state/menu.tsx
@@ -4,7 +4,7 @@ import { MediaPlayerMenubarItem } from "@components/mediaplayer/menu";
 import { CoverMenubarItem } from "@components/cover/menu";
 import { PersonMenubarItem } from "@components/person/menu";
 import { SwitchMenubarItem } from "@components/switch/menu";
-import { CopyToClipboardMenubarItem, MenuBarSubmenu } from "@components/menu";
+import { CopyToClipboardMenubarItem, MenuBarItemConfigureCommand, MenuBarSubmenu } from "@components/menu";
 import { LightMenubarItem } from "@components/light/menu";
 import { WeatherMenubarItem } from "@components/weather/menu";
 import { CameraMenubarItem } from "@components/camera/menu";
@@ -18,6 +18,7 @@ import { VacuumMenubarItem } from "@components/vacuum/menu";
 import { getStateValue, getIcon } from "./utils";
 import { InputButtonMenubarItem } from "@components/input_button/menu";
 import { AutomationMenubarItem } from "@components/automation/menu";
+import { MenuBarExtra } from "@raycast/api";
 
 export function CopyEntityIDToClipboard(props: { state: State }) {
   const s = props.state;
@@ -93,5 +94,29 @@ export function StateMenubarItem(props: { state: State }): JSX.Element | null {
       <CopyEntityIDToClipboard state={s} />
       <CopyEntityStateToClipboardMenubarItem state={s} />
     </MenuBarSubmenu>
+  );
+}
+
+export function MenuBarExtraEntity(props: { state: State | undefined; isLoading?: boolean | undefined }) {
+  const s = props.state;
+  const stateValue = (state: State | undefined) => {
+    if (!state) {
+      return undefined;
+    }
+    const e = state.entity_id;
+    if (e.startsWith("sensor.")) {
+      return getStateValue(state);
+    }
+    return undefined;
+  };
+  return (
+    <MenuBarExtra
+      title={stateValue(s)}
+      icon={s ? getIcon(s) : "entity.png"}
+      tooltip={s ? getFriendlyName(s) : undefined}
+      isLoading={props.isLoading}
+    >
+      <MenuBarItemConfigureCommand />
+    </MenuBarExtra>
   );
 }