+++ extensions/homeassistant/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Home Assistant Changelog
 
+## [Update] - 2024-01-27
+
+- Add Single Entity Menu Command
+
 ## [Update] - 2024-01-19
 
 - Make use of cache for faster loading of entities